added synchronous user vocab saving
I add callbacks to functions to force them to be synchronous so that the
file will be sure to pull the information from the API before it tries
to save it. Also added a checker to see if the vocabulary list has
already been saved. If it has been then call from local storage instead
of making a API call.